% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorts.R
\name{catln}
\alias{catln}
\title{Concatenate and print, supports line breaks from double comma ',,'}
\usage{
catln(..., sep = "")
}
\description{
Concatenate and print, supports line breaks from double comma ',,'
}
\examples{
catln(1, ". This is the first line, and",, 
2, ". this is the second line.")
}
