% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{change_variable}
\alias{change_variable}
\title{Reorder function arguments.}
\usage{
change_variable(f, variables)
}
\arguments{
\item{f}{Function with multiple arguments.}

\item{variables}{Arguments to put first and have their default values removed.}
}
\description{
Reorder function arguments.
}
