% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorts.R
\name{Char}
\alias{Char}
\title{Comfortable character vectors}
\usage{
Char(..., .split = "; ")
}
\arguments{
\item{...}{Anything that can be coerced to character}

\item{.split}{Where to split individual strings.}
}
\description{
Uses a split to separate elements without having to open and close
"" every time. Use \\ before to
}
