% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combn_first}
\alias{combn_first}
\title{Generate the first combinations of n elements, taken m at a time}
\usage{
combn_first(x, m, first, warn = FALSE, FUN = NULL, simplify = TRUE, ...)
}
\arguments{
\item{x}{vector source for combinations, or integer n for x <- seq_len(n).}

\item{m}{number of elements to choose.}

\item{first}{maximum number of combinations to return.}

\item{warn}{logical, whether to warn if the actual number of combinations
does not reach the maximum.}

\item{FUN}{function to be applied to each combination.}

\item{simplify}{logical indicating if the result should be simplified
to an array.}

\item{...}{optionally, further arguments to \code{FUN}.}
}
\description{
Slight modification of \code{\link[utils]{combn}} to stop after
a set number of combinations.
}
\examples{
combn_first(10, 4, first = 5)
}
\seealso{
\code{\link[utils:combn]{utils::combn()}}
}
