% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorts.R
\name{cumsum_na}
\alias{cumsum_na}
\alias{cumprod_na}
\alias{cummax_na}
\alias{cummin_na}
\title{Cumulative Sums, Products, and Extremes for vectors with NA values.}
\usage{
cumsum_na(x, remove = FALSE, replace = TRUE)

cumprod_na(x, remove = FALSE, replace = TRUE)

cummax_na(x, remove = FALSE, replace = TRUE)

cummin_na(x, remove = FALSE, replace = TRUE)
}
\arguments{
\item{x}{Numeric vector.}

\item{remove}{Whether to remove NA values altogether. 
The resulting vector will be of lesser or equal length to x.}

\item{replace}{Whether to replace original NA values in the original vector
with NA in the resulting vector.}
}
\value{
Numeric vector.
}
\description{
Makes it so NA values are neutral.
}
\examples{
x <- rpois(10, 5)
x[c(4, 5, 8)] <- NA
cumsum_na(x)
cumprod_na(x)
cummax_na(x)
cummin_na(x)
}
