% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{derivate}
\alias{derivate}
\title{Analytical Derivative of a Function}
\usage{
derivate(f, arg = formalArgs(f)[1L])
}
\arguments{
\item{f}{Function to derivate. Supports Lambda notation.}

\item{arg}{String, variable to derivate by.}
}
\value{
A new function, which takes the same arguments as f
and returns a single numeric value.
}
\description{
Analytical Derivative of a Function
}
\seealso{
[gradient()]
}
