% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{derivate_num}
\alias{derivate_num}
\title{Numeric Derivative of a Function}
\usage{
derivate_num(f, arg = formalArgs(f)[1L], ...)
}
\arguments{
\item{f}{Function to derivate.}

\item{arg}{String, variable to derivate by.}
}
\value{
A new function.
}
\description{
Numeric Derivative of a Function
}
