% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{distribute}
\alias{distribute}
\title{Distribute an Amount of something Between something.}
\usage{
distribute(amount, between, dist = runif(between), digits = 0)
}
\arguments{
\item{amount}{Numeric, target sum of the values.}

\item{between}{Integer, number of values.}

\item{dist}{Optionally, an initial distribution for how the values are distributed.}

\item{digits}{Digits of the values.}
}
\value{
Numeric vector of length \code{between},
the values of which add up to \code{amount}.
}
\description{
Distribute an Amount of something Between something.
}
