% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand_grid}
\alias{expand_grid}
\title{Expand Grid with repeating support.}
\usage{
expand_grid(..., KEEP.OUT.ATTRS = TRUE, stringsAsFactors = TRUE)
}
\arguments{
\item{...}{Vectors or lists. Use "x \%^\% n" to have "n" columns for the vector "x"}
}
\description{
Expand Grid with repeating support.
}
\examples{
expand_grid(1:3 \%^\% 2, letters[1:2])
}
\seealso{
See \link{expand.grid}, the base version of the function, in which this
one relies on.
}
