% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorts.R
\name{factor}
\alias{factor}
\title{Factor as it should work in base R}
\usage{
factor(
  x,
  levels = unique(x),
  labels = levels,
  exclude = NA,
  ordered = is.ordered(x),
  nmax = NA
)
}
\description{
Levels are ordered as they appear in the data.
}
