% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{for_split}
\alias{for_split}
\title{Split a for loop into expressions, with the looper variable replaced
with each value.}
\usage{
for_split(expr, recursive = TRUE, envir = parent.frame())
for (variable in sequence) expression
}
\arguments{
\item{expr}{Call including all the for loop.}

\item{recursive}{Whether to split for loops inside for loops.}

\item{envir}{Environment where to evaluate the sequence.}
}
\value{
A list of expressions. If the sequence is named, the result copies the names.
If \code{recursive = TRUE}, it will be lists inside of lists.
}
\description{
Useful to split a single expression into multiple, without evaluating anything
but the sequence.
}
\examples{
my_loop <- quote(
    for(i in 1:3) for(j in 1:2)
        a <- a + x[i, j]
)
fsplit <- for_split(my_loop)
fsplit
unlist(fsplit)
}
