% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times.R
\name{format.tempo}
\alias{format.tempo}
\title{Format a tempo object.
with a flexible syntax.}
\usage{
\method{format}{tempo}(
  x,
  format = "\%H:\%M:\%S",
  digits = c(H = 0, M = 0, S = 0, h = 3, m = 3, s = 3),
  trim_zero = FALSE,
  ...
)
}
\arguments{
\item{x}{A tempo object.}

\item{format}{Format to use. See \link{details}.}

\item{digits}{Number of digits to add to \code{\%S}.}

\item{trim_zero}{Whether to trim the first zero. Only works if it's the
first character of the formatted string.}

\item{...}{Ignored.}
}
\value{
A formatted character vector.
}
\description{
Format a tempo object.
with a flexible syntax.
}
\details{
Argument \code{format} must be specified as follows:
\itemize{
\item \code{\%H} : Hours. Can exceed 24.
\item \code{\%M} : Clock Minutes. Cannot exceed 60.
\item \code{\%S} : Clock Seconds. Cannot exceed 60.
\item \code{\%h} : Hours, with 3 decimals by default. Can exceed 24.
\item \code{\%m} : Total Minutes. Can exceed 60.
\item \code{\%s} : Total Seconds. Can exceed 60.
}
}
\examples{

my_tempo <- Tempo(3735)
format(my_tempo)
format(my_tempo, "\%H, \%M, \%S")
format(my_tempo, "It's been \%m minutes")
}
