% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_hms}
\alias{format_hms}
\title{Format hms.}
\usage{
format_hms(x, format = "HMS", digits = 0)
}
\arguments{
\item{x}{\code{hms} object or numeric (in seconds).}

\item{format}{String indicating the format. Maximum 3 characters:
\itemize{
\item H indicates hours.
\item M indicates minutes.
\item S indicates seconds.
\item h indicates hours, will be removed if time is under an hour.
\item m indicates minutes, will be removed if time is under a minute.
}}

\item{digits}{Integer, number of digits to be included for the last value in \code{format}.}
}
\value{
Character vector.
}
\description{
Format hms.
}
