% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{function_list}
\alias{function_list}
\alias{calls_to_functions}
\title{Make a list of functions}
\usage{
function_list(..., .args)

calls_to_functions(exprs, .args)
}
\arguments{
\item{...}{Calls.}

\item{.args}{Arguments shared by all functions.}

\item{exprs}{List of calls.}
}
\value{
List of functions.
}
\description{
Make a list of functions
}
\examples{
my_functions <- function_list(
    .args = c("x", "y"),
    x + mean(y) / z,
    x^2 + sd(y) / z^2,
    sqrt(x) + y + z
)

# Is equivalent to
my_calls <- alist(
    x + mean(y) / z,
    x^2 + sd(y) / z^2,
    sqrt(x) + y + z
)
my_functions <- calls_to_functions(my_calls, c("x", "y"))

# Is equivalent to
my_functions <- list(
    \(x, y) x + mean(y) / z,
    \(x, y) x^2 + sd(y) / z^2,
    \(x, y) sqrt(x) + y + z
)
}
\seealso{
[list_to_function()]
}
