\name{get.current.layer}
\alias{get.current.layer}
\title{Get the ID and name of the Current Layer}
\value{
  A data frame listing the ID and name of the Current
  Layer.
}
\description{
  This function gets the ID and name of the Current Layer.
}
\author{
  Yixuan Qiu (\email{yixuan.qiu@cos.name})
}
