% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{gradient}
\alias{gradient}
\title{Analytical Gradient of a function.}
\usage{
gradient(f, args = formalArgs(f))
}
\arguments{
\item{f}{Function to derivate.}

\item{args}{Character vector, which variables to derivate by.}
}
\value{
A new function which takes the same arguments as f
and returns a named numeric vector.
}
\description{
Analytical Gradient of a function.
}
\seealso{
[derivate()]
}
