% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorts.R
\name{gradient_n}
\alias{gradient_n}
\title{Alias for ColorRampPalette}
\usage{
gradient_n(...)
}
\description{
Uses non named arguments as colors and named arguments are passed on to
\link{colorRamp}
}
\seealso{
\link{colorRamp}
}
