% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{hessian}
\alias{hessian}
\title{THIS IS VERY SLOW AND NEEDS TO BE MADE TWICE AS FAST!!!}
\usage{
hessian(f, args = formalArgs(f))
}
\description{
THIS IS VERY SLOW AND NEEDS TO BE MADE TWICE AS FAST!!!
}
