% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{inside}
\alias{inside}
\title{Get the expression inside an enclosure, e.g. (expr) or {expr}.}
\usage{
inside(expr)
}
\arguments{
\item{expr}{Quoted expression.}
}
\description{
Get the expression inside an enclosure, e.g. (expr) or {expr}.
}
