% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{invert}
\alias{invert}
\title{Invert a function}
\usage{
invert(f, interval, no_root = NaN, arg = formalArgs(f)[1L])
}
\arguments{
\item{f}{Function to invert.}

\item{interval}{Length 2 vector containing the lower and upper bound of the result.}

\item{no_root}{Value to return when there is no root.}

\item{arg}{Argument of f to invert by.}
}
\value{
Inverse function: inv(y, ...) = {x : f(x, ...) = y}
}
\description{
Invert a function
}
