% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_dimensions.R
\name{keep_dim}
\alias{keep_dim}
\alias{unkeep_dim}
\title{Keep dimensions while indexing an object.}
\usage{
keep_dim(x)

unkeep_dim(x)
}
\arguments{
\item{x}{Array.}
}
\value{
The same array, with class "keep_dim"
}
\description{
By default, R will drop the dimensions of an array when a single row or 
column is indexed. \code{x[i, j, ..., drop = TRUE]}.

Objects of class "keep_dim" will override this behaviour, so you don't have
to add \code{drop = FALSE} at every command.

remove "keep_dim" class from an object.
}
\examples{
a <- matrix(1:16, nrow = 4)
a[1, ]
a[, 1]

a <- keep_dim(a)
a[1, ]
a[, 1]
a[, 1, drop = TRUE]
}
