% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landscape_points}
\alias{landscape_points}
\title{Landscape Points}
\format{
A \code{\link[raster]{raster}} object.
}
\source{
Data generated using the \href{https://docs.ropensci.org/NLMR/}{NLMR} package. See Sciaini et al. (2018) for details: \doi{10.1111/2041-210X.13076}.
}
\usage{
landscape_points
}
\description{
Random curd neutral landscape model generated using the \code{NLMR} package.
}
\keyword{datasets}
