% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-package.R
\docType{package}
\name{layer-package}
\alias{layer}
\alias{layer-package}
\title{layer: Tilt your Maps and Turn Them into 'ggplot' Plots}
\description{
Simplifies the process of creating stacked tilted maps, commonly used in
scientific publications to visualize spatial correlations across different
environmental layers.
}
\details{
The package provides a declarative API (\code{tilt_stack()}) to easily build
map stacks, handle spatial transformations, and automatically generate
connecting lines between layers. It also includes high-level animation
support via \code{animate_tilt_stack()}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/marcosci/layer}
  \item \url{https://marcosci.github.io/layer/}
  \item Report bugs at \url{https://github.com/marcosci/layer/issues}
}

}
\author{
\strong{Maintainer}: Egor Kotov \email{kotov.egor@gmail.com} (\href{https://orcid.org/0000-0001-6690-5345}{ORCID}) [contributor]

Authors:
\itemize{
  \item Marco Sciaini \email{marco@kaldera.dev} [copyright holder]
  \item Cédric Scherer \email{cedricphilippscherer@gmail.com}
}

Other contributors:
\itemize{
  \item Josh Erickson \email{joshualerickson@gmail.com} [contributor]
}

}
\keyword{internal}
