% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorts.R
\name{len0}
\alias{len0}
\title{Check if an argument is missing, NULL, or of length 0.}
\usage{
len0(x)
}
\arguments{
\item{x}{Usually a function argument, can also be any object.}
}
\description{
Check if an argument is missing, NULL, or of length 0.
}
