% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorts.R
\name{length_unique}
\alias{length_unique}
\title{Number of unique values.}
\usage{
length_unique(x, include_na = TRUE)
}
\arguments{
\item{x}{Array or factor.}

\item{include_na}{Logical, whether to count missing values as a different unique value.
By default it is \code{TRUE} so that it gives the same result as \code{length(unique(x))}.
If \code{FALSE}, it is equivalent to \verb{length(unique(x[!is.na(x)])} but much faster.}
}
\value{
Integer indicating the number of unique values.
}
\description{
If \code{include_na = FALSE}, always gives the same result as
\code{length(unique(x))}, but much faster.
}
\examples{
x <- rpois(100, 3.5)
length_unique(x)
# length(unique(x))

x[1:3] <- NA
length_unique(x)
# length(unique(x))
length_unique(x, include_na = FALSE)
# length(unique(x[!is.na(x)])
}
