% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorts.R
\name{map_range}
\alias{map_range}
\title{Map a numeric vector from one range to another}
\usage{
map_range(x, from = range(x), to = 0:1, clamp = FALSE)
}
\arguments{
\item{x}{Numbers to map.}

\item{from}{Range to map from.}

\item{to}{Range to map to.}

\item{clamp}{Whether to force numbers within the range.}
}
\value{
A vector of the same length, with numbers
}
\description{
Map a numeric vector from one range to another
}
\examples{
map_range((1:10)^2, to = 2:8) \%>\% barplot()
map_range((1:10)^2, from = 0:50, to = 2:8, clamp = TRUE) \%>\% barplot()
}
