% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{percent}
\alias{percent}
\title{Object of class 'percent'}
\usage{
percent(x)
}
\arguments{
\item{x}{Numeric vector or array}
}
\description{
Usefull for probabilities
}
\examples{
m <- matrix(runif(36), 6, 6) \%>\% percent()
print(m)
format(m, digits = 2L)
}
