% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorts.R
\name{pipe_if}
\alias{pipe_if}
\title{Comfortable control flow inside the pipe}
\usage{
pipe_if(.data, .condition, .if, .else = .data)
}
\arguments{
\item{.data}{Input object.}

\item{.condition}{Condition.}

\item{.if}{Statement if condition is TRUE.}

\item{.else}{Statement if condition is FALSE.}
}
\description{
Allows the use of the dot anywhere. By default, if the condition is not met
returns itself. Parameters \code{.condition, .if, .else}  can be functions,
in which case they will be evaluated with \code{.data} as a parameter.
}
\examples{
10 \%>\% pipe_if(TRUE, . + 2)
10 \%>\% pipe_if(is.numeric, as.character)
}
