% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilt_stack.R
\name{plot_tilt_stack}
\alias{plot_tilt_stack}
\title{Plot a tilt stack}
\usage{
plot_tilt_stack(stack)
}
\arguments{
\item{stack}{A \code{tilt_stack} object.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot a tilt stack
}
\examples{
\donttest{
stack <- tilt_stack(x_shift_step = 25, y_shift_step = 50)
stack <- tilt_layer(stack, landscape_1, palette = "bilbao")
stack <- tilt_layer(stack, landscape_2, palette = "mako")
stack <- tilt_connector(stack, landscape_points)
plot_tilt_stack(stack)
}
}
