% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilt_stack.R
\name{print.tilt_stack}
\alias{print.tilt_stack}
\title{Print a tilt stack}
\usage{
\method{print}{tilt_stack}(x, ...)
}
\arguments{
\item{x}{A \code{tilt_stack} object.}

\item{...}{Additional arguments passed to \code{print}.}
}
\value{
The original \code{tilt_stack} object, invisibly.
}
\description{
Print a tilt stack
}
