% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pyramid}
\alias{pyramid}
\title{Negate values to make a pyramid plot.}
\usage{
pyramid(.data, values, by, level = fact[1L])
}
\arguments{
\item{.data}{Data frame or tibble.}

\item{values}{Columns with the values to negate.}

\item{by}{Column, preferably a factor, to negate values by.}

\item{level}{Level in the factor to have it's values negated.}
}
\value{
An object of the same class as .data
}
\description{
Negate values to make a pyramid plot.
}
