% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{rep_array}
\alias{rep_array}
\title{Replicate an Array}
\usage{
rep_array(x, margin = 1, times = 1, each = 1)
}
\arguments{
\item{x}{Array object.}

\item{margin}{Margin to stack by. 1 means rows, 2 means columns. The order of
the margin matters because it is done iteratively.}
}
\value{
Object of the same class as x.
}
\description{
Stack an array or dataframe by replicating it through
one or more margins.
}
