% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorts.R
\name{return_at}
\alias{return_at}
\title{Call "return" from another function's execution environment.}
\usage{
return_at(x, n = 1L, envir)
}
\arguments{
\item{x}{Object to return.}

\item{n}{Number of generations to go back. By default, it
executes \code{return(x)} at the function which calls it.
See \link{examples}. Ignored when 'envir' is specified.}

\item{envir}{Alternatively, the environment where \code{return(x)} is to be
executed. When specified, 'n' is ignored.}
}
\value{
Invisibly returns 'x'
}
\description{
By, default, evaluates in
}
\examples{
f <- function() {
    return_at(1)
}

g <- function() {
    f()
    return(0)
}

result <- g()
result
}
