% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{rev_array}
\alias{rev_array}
\title{Reverse an array by one or more margins.}
\usage{
rev_array(x, margin = 1)
}
\arguments{
\item{x}{Array object.}

\item{margin}{Margin to reverse by. 1 means rows, 2 means columns. The order of
the margin matters because it is done iteratively.}
}
\value{
Object of the same class as x.
}
\description{
Reverse an array by one or more margins.
}
