% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_sum}
\alias{round_sum}
\title{Round a vector such that their rounded sum stays intact.}
\usage{
round_sum(x, digits = 0)
}
\arguments{
\item{x}{Numeric Vector}

\item{digits}{Number of decimal places. To mak}
}
\value{
Rounded vector with attributes:
Threshold: indicating at which point numbers are rounded to their ceiling.
Relative Threshold: the above in relation to the digits, to be compared with 0.5.
Bias: mean difference between the result and the default rounding with the
same digits.
}
\description{
Useful for percentatges.
}
\examples{
x <- rexp(20)
x <- 100 * x / sum(x)
x
round_sum(x, 2)
sum(round_sum(x, 2))
}
