% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorts.R
\name{sample}
\alias{sample}
\title{Overwrites default behavior of sample to avoid silent errors.}
\usage{
sample(x, size, replace = FALSE, prob = NULL)
}
\description{
When x is a length 1 integer, it won't use sample.int. See
the base definition of sample.
}
\seealso{
\code{\link[base:sample]{base::sample()}}
}
