\name{set.current.layer}
\alias{set.current.layer}
\title{Set a layer as the Current Layer}
\usage{
set.current.layer(id)
}
\arguments{
  \item{id}{the unique ID number of the layer.}
}
\value{
  A data frame listing the ID and name of this layer.
}
\description{
  This function sets a layer given by its ID as the Current
  Layer.
}
\author{
  Yixuan Qiu (\email{yixuan.qiu@cos.name})
}
