% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorts.R
\name{set}
\alias{set}
\title{Perform set operations elegantly.}
\usage{
set(expr)
}
\arguments{
\item{expr}{Expression}
}
\value{
A set.
}
\description{
Use \code{|} for set unions, \code{&} for intersection, and \code{~} for
difference.
}
\examples{
A <- 1:10
B <- 5:12
C <- 2:3
set(A | B ~ C)
}
