% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_defaults}
\alias{set_defaults}
\alias{reset_defaults}
\title{Change a function's default parameters}
\usage{
set_defaults(fun, ..., reassign = TRUE, envir = globalenv(), warn = TRUE)

reset_defaults(
  fun,
  envir = parent.env(globalenv()),
  reassign = TRUE,
  reassign_envir = globalenv()
)
}
\arguments{
\item{fun}{A function,}

\item{...}{Named values, accepts injection.}

\item{envir}{Where to start looking the old function with the
default default parameters.}
}
\value{
The modified function.
}
\description{
Change a function's default parameters
}
\details{
If you overwrite a function, you won't be able to retrieve the old
defaults. You can save them before overwriting by using \link{formals}
}
\examples{
use_names <- FALSE
diag <- set_defaults(diag, nrow = 3, ncol = nrow / 2, names = !!use_names)
old_diag <- reset_defaults(diag)

}
