% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{stack_vectors}
\alias{stack_vectors}
\title{Stack multiple vectors of different length.}
\usage{
stack_vectors(..., .names_to = "name", .values_to = "value")
}
\arguments{
\item{...}{\code{name1 = vector1, name2 = vector2, ...} \cr
The names are sent to the column \code{.names_to}
as a factor, and the vectors are stacked on the columns \code{.values_to}.}

\item{.names_to}{Name of the factor column.}

\item{.values_to}{Name of the value column.}
}
\value{
A tibble where the name column is a factor and the value column are the stacked vectors.
}
\description{
Stack multiple vectors of different length.
}
\examples{
stack_vectors(a = runif(6), b = rpois(4, 2))
stack_vectors(a = runif(6), b = rpois(4, 2),
    .names_to = "category",
    .values_to = "number")
}
