% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_replace_fun}
\alias{str_replace_fun}
\title{Replace matches in a string with a transformed match.}
\usage{
str_replace_fun(string, pattern, fun, ...)
}
\arguments{
\item{string}{A character vector.}

\item{pattern}{Pattern to match against.}

\item{fun}{Function to transform matches.}

\item{...}{Arguments to be passed on to fun.}
}
\value{
The transformed string.
}
\description{
Replace matches in a string with a transformed match.
}
