% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_split_index}
\alias{str_split_index}
\title{Split strings by the character index}
\usage{
str_split_index(string, index, where = "left")
}
\arguments{
\item{string}{String to be split.}

\item{index}{Index at which to split.}

\item{where}{Where to insert the character at the indexes. Possible values
are 'left', 'right', 'middle', and 'remove'.}
}
\value{
The split string as a vector.
}
\description{
Split strings by the character index
}
