% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_split_keep}
\alias{str_split_keep}
\title{Split a string while keeping the pattern.}
\usage{
str_split_keep(string, pattern, where = "left")
}
\arguments{
\item{string}{The string to split.}

\item{pattern}{Pattern to split with.}

\item{where}{Where to attach pattern. Can be 'left', 'right' or 'middle',
which will keep the pattern as a separate string.}
}
\value{
A list of character vectors.
}
\description{
Split a string while keeping the pattern.
}
