% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{such_that}
\alias{such_that}
\title{Find the value of a variable such that an equality is true}
\usage{
such_that(variable, equality, interval)
}
\arguments{
\item{variable}{Variable to find the value of.}

\item{equality}{Equality expression.}

\item{interval}{Interval where the value can be.}
}
\value{
The value of the variable if it's in the interval, an error otherwise.
}
\description{
Find the value of a variable such that an equality is true
}
\examples{
x_hat <- x |> such_that(exp(x) == 5*x^2, interval = c(1, 10))
}
