% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{through}
\alias{through}
\title{Catch attempt to use through outside of traverse or nested inside another through.}
\usage{
through(x)
}
\arguments{
\item{x}{Whatever}
}
\value{
An error.
}
\description{
Catch attempt to use through outside of traverse or nested inside another through.
}
