% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilt_animate.R
\name{tilt_match}
\alias{tilt_match}
\title{Match data to a layer in a tilt stack}
\usage{
tilt_match(data, stack, layer = 1)
}
\arguments{
\item{data}{An \code{sf} or raster object to transform.}

\item{stack}{A \code{tilt_stack} object.}

\item{layer}{The index of the layer in the stack to match.}
}
\value{
A transformed \code{sf} object.
}
\description{
This utility takes an \code{sf} or raster object and transforms it to perfectly align
with the spatial perspective of a specific layer in an existing \code{tilt_stack}.
}
\examples{
\donttest{
stack <- tilt_stack() |> tilt_layer(landscape_1)
points_tilted <- tilt_match(landscape_points, stack, layer = 1)
}
}
