% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{translate}
\alias{translate}
\title{Replace matching values with respective counterparts.}
\usage{
translate(x, old, new)
}
\arguments{
\item{x}{The vector to be translated.}

\item{old}{Old values to be replaced.}

\item{new}{Replacement for the old values.}
}
\value{
The translated vector.
}
\description{
Replace matching values with respective counterparts.
}
