% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{traverse}
\alias{traverse}
\title{Traverse through one or more vector or lists in parallel.}
\usage{
traverse(expr)
}
\arguments{
\item{expr}{Expression to evaluate.}
}
\value{
A list.
}
\description{
Super cool alternative to [base::mapply()].
Use \code{through(vector)} to tell the apply the expression to each of the values of \code{vector}.
You can use multiple \code{through()} statements to traverse the multiple vectors in parallel.
It is NOT possible to nest a \code{through()} statement inside another, and will return an error.
}
\examples{
x <- list(1:3, 4:5, 6:9)
y <- c(1, 2, 4)
traverse( through(x) + through(y) )
traverse( sum(through(x), through(y)) )
traverse( letters[through(x)] )
traverse( rep(through(letters), times = through(1:26)) )
}
