% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorts.R
\name{Vec}
\alias{Vec}
\title{Create a vector with trailing arguments.}
\usage{
Vec(...)
}
\arguments{
\item{...}{Elements.}
}
\value{
A vector
}
\description{
Create a vector with trailing arguments.
}
\examples{
Vec(1, 2, 3,)
c(1, 2, 3,)
}
