% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorts.R
\name{within_list}
\alias{within_list}
\title{Alias for within(list(), {})}
\usage{
within_list(expr)
}
\arguments{
\item{expr}{{}}
}
\value{
A list environment, with expression executed.
}
\description{
Alias for within(list(), {})
}
