% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorts.R
\name{wrap}
\alias{wrap}
\title{Wrap a numeric vector between a minimum and a maximum,
as understood in Blender}
\usage{
wrap(x, min, max)
}
\arguments{
\item{x}{Numbers to wrap.}

\item{min}{Minimum, inclusive.}

\item{max}{Maximum, not inclusive.}
}
\description{
Simply calculates  \code{min + (x - min) \%\% (max - min)}
}
\details{
When wrapping integers, it is sometimes useful for the Maximum to be
inclusive, instead of the Minimum.
To achieve this, just add +0.1 to both Minimum and Maximum.
}
