% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_png}
\alias{write_png}
\title{Save a plot as a png.}
\usage{
write_png(
  p,
  filename,
  asp = 16/9,
  quality = 2.5,
  visual_size = 1,
  plot_it = TRUE,
  gg = TRUE
)
}
\arguments{
\item{p}{Either a ggplot object or an expression that plots something.}

\item{filename}{The name of the file where the plot will be saved.}

\item{asp}{Aspect ratio of the plot.}

\item{quality}{Quality of the image. Does not affect size of elements relative
to the plot size.}

\item{visual_size}{Visual size of plot elements like text, line width...}

\item{plot_it}{Whether to plot it in the active graphical device after connection
with .png file is closed.}

\item{gg}{TRUE if 'p' is a ggplot object, FALSE if it's an expression.}
}
\description{
Save a plot as a png.
}
