% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorts.R
\name{catln}
\alias{catln}
\title{Concatenate and print, supports line breaks from double comma ',,'}
\usage{
catln(..., sep = "")
}
\description{
Concatenate and print, supports line breaks from double comma ',,'
}
\examples{
catln("The answer to all is", 42,, "And 2 + 3 = ", 5)
}
