% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{derivate}
\alias{derivate}
\title{Analytical Derivative of a Function}
\usage{
derivate(f, arg = formalArgs(f)[1L])
}
\arguments{
\item{f}{Function to derivate.}

\item{arg}{Variable to derivate by.}
}
\value{
A new function.
}
\description{
Analytical Derivative of a Function
}
