% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times.R
\name{format.tempo}
\alias{format.tempo}
\title{Format a tempo object.
with a flexible syntax.}
\usage{
\method{format}{tempo}(x, format = "\%H:\%M:\%S", digits = 0L, trim_zero = FALSE, ...)
}
\arguments{
\item{x}{A tempo object.}

\item{format}{Format to use. See \link{details}.}

\item{digits}{Number of digits to add to \code{\%S}.}

\item{trim_zero}{Whether to trim the first zero. Only works if it's the
first character of the formatted string.}

\item{...}{Ignored.}
}
\value{
A formatted character vector.
}
\description{
Format a tempo object.
with a flexible syntax.
}
\details{
Argument \code{format} must be specified as follows:

\code{\%H}  Hours. Can exceed 24.
\code{\%M}  Clock Minutes. Can't exceed 60.
\code{\%S}  Clock Seconds. Can't exceed 60.
\code{\%h}  Hours with decimals. Can exceed 24.
\code{\%m}  Total Minutes with decimals. Can exceed 60.
\code{\%s}  Total Seconds with decimals. Can exceed 60.
}
\examples{

my_tempo <- Tempo(3735)
format(my_tempo)
format(my_tempo, "\%H, \%M, \%S")
format(my_tempo, "It's been \%m minutes")
}
