% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landscape_1}
\alias{landscape_1}
\title{Landscape 1.}
\format{
A \code{\link[raster]{raster}} object.
}
\source{
NLMR package
}
\usage{
landscape_1
}
\description{
Random curd neutral landscape model.
}
\examples{
if (requireNamespace("raster", quietly = TRUE)) {
  raster::plot(landscape_1)
}
}
\keyword{datasets}
