% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landscape_1}
\alias{landscape_1}
\title{Landscape 1}
\format{
A \code{\link[raster]{raster}} object.
}
\source{
Data generated using the \href{https://ropensci.github.io/NLMR/}{NLMR} package. See Sciaini et al. (2018) for details: \doi{10.1111/2041-210X.13076}.
}
\usage{
landscape_1
}
\description{
Random curd neutral landscape model generated using the \code{NLMR} package.
}
\keyword{datasets}
