% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landscape_2}
\alias{landscape_2}
\title{Landscape 2.}
\format{
raster object
}
\source{
NLMR package
}
\usage{
landscape_2
}
\description{
Fractional Brownian motion neutral landscape model.
}
\keyword{datasets}
