% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landscape_3}
\alias{landscape_3}
\title{Landscape 3.}
\format{
raster object
}
\source{
NLMR package
}
\usage{
landscape_3
}
\description{
Distance gradient neutral landscape model.
}
\keyword{datasets}
