% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landscape_3}
\alias{landscape_3}
\title{Landscape 3}
\format{
A \code{\link[raster]{raster}} object.
}
\source{
Data generated using the \code{\link[NLMR]{NLMR}} package. See Sciaini et al. (2018) for details: \doi{10.1111/2041-210X.13076}.
}
\usage{
landscape_3
}
\description{
Distance gradient neutral landscape model generated using the \code{NLMR} package.
}
\keyword{datasets}
