% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landscape_points}
\alias{landscape_points}
\title{Landscape Points.}
\format{
raster object
}
\source{
NLMR package
}
\usage{
landscape_points
}
\description{
Random curd neutral landscape model.
}
\keyword{datasets}
