% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-package.R
\docType{package}
\name{layer-package}
\alias{layer}
\alias{layer-package}
\title{layer: Tilt your Maps and Turn Them into 'ggplot' Plots}
\description{
Simplifies the whole process of creating stacked tilted maps, that are often used in scientific publications to show different environmental layers for a geographical region. Tilting maps and layering them allows to easily draw visual correlations between these environmental layers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/marcosci/layer}
  \item \url{https://marcosci.github.io/layer/}
  \item Report bugs at \url{https://github.com/marcosci/layer/issues}
}

}
\author{
\strong{Maintainer}: Egor Kotov \email{kotov.egor@gmail.com} (\href{https://orcid.org/0000-0001-6690-5345}{ORCID}) [contributor]

Authors:
\itemize{
  \item Marco Sciaini \email{marco@kaldera.dev} [copyright holder]
  \item Cédric Scherer \email{cedricphilippscherer@gmail.com}
}

Other contributors:
\itemize{
  \item Josh Erickson \email{joshualerickson@gmail.com} [contributor]
}

}
\keyword{internal}
