% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{minimize}
\alias{minimize}
\alias{maximize}
\title{Optimize a multi-variate function.}
\usage{
minimize(f, init, ...)

maximize(f, init, ...)
}
\arguments{
\item{f}{Function to optimise.}

\item{init}{Initial values for the variable.}

\item{...}{Arguments passed on to  \link[stats]{optim}.}
}
\value{
List with the \code{solution} and \code{objective_value}.
}
\description{
Optimize a multi-variate function.
}
