% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matrix_byrow}
\alias{matrix_byrow}
\title{Convinient matrix definition.}
\usage{
matrix_byrow(..., .fill, .nrow, .ncol)
}
\description{
Supports ',,' for row breaks.
}
\seealso{
\link{tibble::frame_matrix}
}
