% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tiltmaps.R
\name{plot_tiltedmaps}
\alias{plot_tiltedmaps}
\title{Tilt raster and sf data}
\usage{
plot_tiltedmaps(map_list)
}
\arguments{
\item{data}{sf or terra/stars/raster object.}
}
\value{
ggplot
}
\description{
Takes tilted maps and plots them with ggplot.
}
\examples{
# tilt data
tilt_landscape_1 <- tilt_map(landscape_1)
tilt_landscape_2 <- tilt_map(landscape_2, x_shift = 50, y_shift = 50)

# put in list
map_list <- list(tilt_landscape_1, tilt_landscape_2)

# plot
plot_tiltedmaps(map_list)
}
