% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilt_stack.R
\name{tilt_layer}
\alias{tilt_layer}
\title{Add a layer to a tilt stack}
\usage{
tilt_layer(
  stack,
  data,
  x_shift = NULL,
  y_shift = NULL,
  x_stretch = NULL,
  y_stretch = NULL,
  x_tilt = NULL,
  y_tilt = NULL,
  angle_rotate = NULL,
  fill = "value",
  color = "grey50",
  palette = "viridis",
  direction = 1,
  begin = 0,
  end = 1,
  alpha = 1,
  size = NULL,
  label = NA,
  label_x = NULL,
  label_y = NULL,
  label_color = NULL,
  label_size = NULL
)
}
\arguments{
\item{stack}{A \code{tilt_stack} object.}

\item{data}{An \code{sf} or raster object.}

\item{x_shift}{Shift in x dimension. If \code{NULL}, inherits from stack defaults.}

\item{y_shift}{Shift in y dimension. If \code{NULL}, inherits from stack defaults.}

\item{x_stretch}{Stretch in x dimension. If \code{NULL}, inherits from stack defaults.}

\item{y_stretch}{Stretch in y dimension. If \code{NULL}, inherits from stack defaults.}

\item{x_tilt}{Tilt in x dimension. If \code{NULL}, inherits from stack defaults.}

\item{y_tilt}{Tilt in y dimension. If \code{NULL}, inherits from stack defaults.}

\item{angle_rotate}{Rotation angle. If \code{NULL}, inherits from stack defaults.}

\item{fill}{Column name to use for fill and color aesthetics. If \code{NA} (or if the column is not found in the data), a static color is used. Defaults to \code{"value"}.}

\item{color}{A single color string used if \code{fill} is \code{NA} or not found.}

\item{palette}{Palette name from the \link[viridis]{viridis} or \link[scico]{scico} package.}

\item{direction}{Direction for the color palette (1 or -1).}

\item{begin}{Start of interval for palette.}

\item{end}{End of interval for palette.}

\item{alpha}{Opacity of the layer.}

\item{size}{Size of the points or lines in the layer. If \code{NULL}, defaults to \code{0.01} for the first layer and \code{0.5} for others.}

\item{label}{Optional text label to annotate the layer.}

\item{label_x}{X coordinate for the label. If \code{NULL}, inherits from stack defaults.}

\item{label_y}{Y coordinate for the label. If \code{NULL}, inherits from stack defaults or \code{y_shift}.}

\item{label_color}{Color for the label. If \code{NULL}, inherits from stack defaults.}

\item{label_size}{Size for the label. If \code{NULL}, inherits from stack defaults.}
}
\value{
The updated \code{tilt_stack}.
}
\description{
Add a layer to a tilt stack
}
