% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilt_stack.R
\name{tilt_stack}
\alias{tilt_stack}
\title{Initialize a tilt stack}
\usage{
tilt_stack(
  x_shift_step = 0,
  y_shift_step = 0,
  x_stretch = 2,
  y_stretch = 1.2,
  x_tilt = 0,
  y_tilt = 1,
  angle_rotate = pi/20,
  label_x = NULL,
  label_y = NULL,
  label_color = "grey40",
  label_size = 4
)
}
\arguments{
\item{x_shift_step}{Default shift in x dimension per layer.}

\item{y_shift_step}{Default shift in y dimension per layer.}

\item{x_stretch}{Default stretch in x dimension.}

\item{y_stretch}{Default stretch in y dimension.}

\item{x_tilt}{Default tilt in x dimension.}

\item{y_tilt}{Default tilt in y dimension.}

\item{angle_rotate}{Default rotation angle.}

\item{label_x}{Default X coordinate for layer labels.}

\item{label_y}{Default Y coordinate for layer labels. If \code{NULL}, defaults to the layer's \code{y_shift}.}

\item{label_color}{Default color for layer labels.}

\item{label_size}{Default size for layer labels.}
}
\value{
An object of class \code{tilt_stack}.
}
\description{
Initialize a tilt stack
}
\examples{
tilt_stack()
}
